sub util 함수 통합
