+++ src/app/(sub)/_componenets/MealPlanItem.tsx
@@ -1,11 +1,11 @@
 import { MealType, MealItemType } from "@/app/(sub)/mealplan/mealType";
 import { MEAL_ITEM_TITLE } from "@/constants/mainConstants";
 import useMealPlanStore from "@/store/mealPlanStore";
-import { toFixeNumberTwo } from "@/utils/mealplanFns";
 import { BsSearch } from "@react-icons/all-files/bs/BsSearch";
 import { BsX } from "@react-icons/all-files/bs/BsX";
 import { useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
+import { toFixeNumberTwo } from "../util";
 
 const MealPlanItem = ({
   data,
