+++ src/app/(sub)/util.ts
@@ -1,3 +1,40 @@
+// calendar fns
+const isLeapYear = (year: number) => {
+  return (year % 4 === 0 && year % 100 !== 0) || year % 400 === 0;
+};
+export const getDaysInMonth = (year: number, month: number) => {
+  return [
+    31,
+    isLeapYear(year) ? 29 : 28,
+    31,
+    30,
+    31,
+    30,
+    31,
+    31,
+    30,
+    31,
+    30,
+    31,
+  ][month - 1];
+};
+
+export const formattedNumber = (num: number) => {
+  return num < 10 ? `0${num}` : String(num);
+};
+
+// mealplan fns
+
+export const toKRLocaleString = (date: Date) => {
+  return date.toLocaleString("ko-KR", {
+    year: "numeric",
+    month: "2-digit",
+    day: "2-digit",
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+};
+
 export function menuTypeTransform(value: string): number;
 export function menuTypeTransform(value: number): string;
 export function menuTypeTransform(value: string | number): number | string {
@@ -61,3 +98,10 @@ export function reCalcMenuSpec(
   }
   return num;
 }
+
+// input fns
+
+export const checkSpecialCharacters = (text: string) => {
+  const expText = /[%=*><]/;
+  return expText.test(text);
+};
