--- src/utils/inputFns.ts
@@ -1,4 +1,0 @@
-export const checkSpecialCharacters = (text: string) => {
-  const expText = /[%=*><]/;
-  return expText.test(text);
-};
